refactor(port): extract Protocol.GUI module for structured chrome encoding
Move all GUI chrome encoding functions (tab bar, file tree, which-key,
completion, breadcrumb, status bar, picker, agent chat, theme) from
Protocol to Protocol.GUI. Theme color slot mapping extracted to
Theme.Slots as a pure data-mapping module that returns [{slot_id, rgb}]
pairs. Binary serialization stays in Protocol.GUI where it belongs.

Protocol.ex: 1,571 → 925 lines (41% reduction), now focused on
TUI cell-grid encoding and input event decoding.

New modules:
- lib/minga/port/protocol/gui.ex (588 lines)
- lib/minga/theme/slots.ex (157 lines)
- test/minga/theme/slots_test.exs (10 tests)

Closes #690
